Смена местоположения файлов не имеющих отношения к gitter боту
