+++ app/Domains/Bot/Middlewares/GoogleSearchMiddleware.php
@@ -2,7 +2,7 @@
 namespace Domains\Bot\Middlewares;
 
 use Domains\Message;
-use Interfaces\Gitter\Middleware\MiddlewareInterface;
+use Domains\Middleware\MiddlewareInterface;
 
 /**
  * Class GoogleSearchMiddleware
