+++ app/Domains/Bot/Middlewares/KarmaRenderMiddleware.php
@@ -2,7 +2,7 @@
 namespace Domains\Bot\Middlewares;
 
 use Domains\Message;
-use Interfaces\Gitter\Middleware\MiddlewareInterface;
+use Domains\Middleware\MiddlewareInterface;
 
 /**
  * Проверяет слово "карма" и выводит статус
