+++ app/Domains/Subscriber/Storage.php
@@ -9,10 +9,10 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace Interfaces\Gitter\Subscriber;
+namespace Domains\Subscriber;
 
 use Illuminate\Contracts\Container\Container;
-use Interfaces\Gitter\Room\RoomInterface;
+use Domains\Room\RoomInterface;
 
 /**
  * Class Storage
